qboot: Move inline load and store routines to memaccess.h
Move lduw_p, ldl_p, stw_p, stl_p from linuxboot.c to
include/memaccess.h so they can be reused more easily.
